// src/plugins/organizations/gql/generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  JSON: { input: any; output: any; }
  Money: { input: number; output: number; }
  Point: { input: any; output: any; }
  Upload: { input: any; output: any; }
};
// ...
export type CreateOrganizationInput = {
  code: Scalars['String']['input'];
  description: Scalars['String']['input'];
  email: Scalars['String']['input'];
  enabled: Scalars['Boolean']['input'];
  name: Scalars['String']['input'];
  ownerId: Scalars['ID']['input'];
};
// ...
export type UpdateOrganizationInput = {
  addressesId?: InputMaybe<Array<Scalars['ID']['input']>>;
  affiliatedWidthId?: InputMaybe<Array<Scalars['ID']['input']>>;
  branchesId?: InputMaybe<Array<Scalars['ID']['input']>>;
  code?: InputMaybe<Scalars['String']['input']>;
  collaboratorsId?: InputMaybe<Array<Scalars['ID']['input']>>;
  defaultAddressId?: InputMaybe<Scalars['ID']['input']>;
  description?: InputMaybe<Scalars['String']['input']>;
  email?: InputMaybe<Scalars['String']['input']>;
  enabled?: InputMaybe<Scalars['Boolean']['input']>;
  id: Scalars['ID']['input'];
  linksRRSS?: InputMaybe<Array<Scalars['String']['input']>>;
  name?: InputMaybe<Scalars['String']['input']>;
  ownerId?: InputMaybe<Scalars['ID']['input']>;
  typeId?: InputMaybe<Scalars['ID']['input']>;
};